requirements.txt
websockets==12.0
fastapi
uvicorn[standard]
cryptography
pyjwt>=2.8.0
python-multipart
pydantic-settings>=2.0.0
python-dotenv
paramiko>=3.4.0
aiofiles
httpx

# Celery worker dependencies for PoC
celery>=5.2
requests
boto3

home-console-sdk
